+++ node/src/Transaction.ts
@@ -48,6 +48,7 @@ import {
     SortClusterOptions,
     SortOptions,
     StreamAddOptions,
+    StreamRangeBound,
     StreamReadOptions,
     StreamTrimOptions,
     ZAddOptions,
@@ -183,6 +184,7 @@ import {
     createXAdd,
     createXDel,
     createXLen,
+    createXRange,
     createXRead,
     createXTrim,
     createZAdd,
@@ -2074,6 +2076,35 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public time(): T {
         return this.addAndReturn(createTime());
+    }
+
+    /**
+     * Returns stream entries matching a given range of IDs.
+     *
+     * See https://valkey.io/commands/xrange for more details.
+     *
+     * @param key - The key of the stream.
+     * @param start - The starting stream ID bound for the range.
+     *     - Use `id` to specify a stream ID.
+     *     - Use `exclusive: "("` to specify an exclusive bounded stream ID.
+     *     - Use `-` to start with the minimum available ID.
+     * @param end - The ending stream ID bound for the range.
+     *     - Use `id` to specify a stream ID.
+     *     - Use `exclusive: "("` to specify an exclusive bounded stream ID.
+     *     - Use `+` to end with the maximum available ID.
+     * @param count - An optional argument specifying the maximum count of stream entries to return.
+     *     If `count` is not provided, all stream entries in the range will be returned.
+     *
+     * Command Response - A mapping of stream IDs to stream entry data, where entry data is a
+     *     list of pairings with format `[[field, entry], [field, entry], ...]`.
+     */
+    public xrange(
+        key: string,
+        start: StreamRangeBound,
+        end: StreamRangeBound,
+        count?: number,
+    ): T {
+        return this.addAndReturn(createXRange(key, start, end, count));
     }
 
     /**